Moving to class per file model.
